updated java samples

+++ samples/client/petstore/java/jersey2-java8/src/test/java/io/swagger/client/api/PetApiTest.java
@@ -67,7 +67,7 @@ public class PetApiTest {
     @Test
     public void deletePetTest() throws ApiException {
         
-        Integer petId = null;
+        Long petId = null;
         
         String apiKey = null;
         
@@ -122,8 +122,8 @@ public class PetApiTest {
      */
     @Test
     public void getPetByIdTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         Pet response = api.getPetById(petId);
 
@@ -158,8 +158,8 @@ public class PetApiTest {
      */
     @Test
     public void updatePetWithFormTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         String name = null;
 
@@ -180,8 +180,8 @@ public class PetApiTest {
      */
     @Test
     public void uploadFileTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         String additionalMetadata = null;
